move default Strategy to a new file & rename
